move: `navigationBar.tsx` をappディレクトリ直下に移動
